third_party/WebKit/LayoutTests/fast/images/image-change-without-resize-shouldnt-layout.html

<!DOCTYPE html>
<title>This test ensures that when an image changes, but doesn't change size (like in an Animated GIF), we don't force a Layout.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div style="float: left;">
<img style="max-width: 100%" src="../replaced/resources/1x1-blue.png">
</div>
<script>
//TODO(siva.gunturi): Use async_test and use asserts on image.onload and window.onload.
test(function() {
    document.documentElement.offsetTop;
    var img = document.querySelector("img");
    img.src = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAMAAAAoyzS7AAAABGdBTUEAAK/INwWK6QAAABl0RVh0U29mdHdhcmUAQWRvYmUgSW1hZ2VSZWFkeXHJZTwAAAAGUExURf8AAAAAAEGjEgMAAAAMSURBVHjaYmAACDAAAAIAAU9tWeEAAAAASUVORK5CYII=";
    assert_equals(internals.needsLayoutCount(), 0);
});
</script>
